OPD Controller Changes

+++ src/main/java/com/hims/entity/DgOrderHd.java
@@ -59,13 +59,13 @@ public class DgOrderHd {
     private String otherInvestigation;
 
     @Column(name = "hospital_id")
-    private int hospitalId ;
+    private Long hospitalId ;
 
     @Column(name = "prescribed_by")
     private int prescribedBy ;
 
     @Column(name = "department_id")
-    private int departmentId ;
+    private Long departmentId ;
 
     @Column(name = "investigation_request_no")
     private int investigationRequestNo;
